PCA component separation
